feat: add week view to calendar page refactor: split calendar page into smaller components refactor: renamed user settings context file
