+++ src/components/settings/social-media-settings.tsx
@@ -10,7 +10,7 @@ import toast from "react-hot-toast"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form"
-import { useUserSettings } from "@/contexts/user-settings-context"
+import { useUserSettings } from "@/contexts/UserSettingsContext"
 import api from "@/services/api-service"
 import { BusinessSocialResponse } from "@/types/response"
 import { useEffect } from "react"
